Course: float instead of double in all lesson code and prose
Students leave this course for Unity, where every position, speed and
angle is a float. Meeting double here and Mathf.Sqrt there is a needless
translation step, so the course now writes float everywhere - in the code
students read and edit, and in the prose that names the type.

The convention, now written down in AUTHORING's Style section: a decimal
literal carries the f suffix (0.5f, 4.5f, 7.0f), a whole number does not
(float speed = 3;). The implicit int-to-float conversion covers the
second, and a bare 20 reads better to a beginner than 20f. List<double>
becomes List<float>; Math.Sqrt and Math.Abs become MathF.Sqrt and
MathF.Abs, which return float, so no cast is needed and none is taught.

Lesson 2 now introduces float as THE decimal type, explains the f in one
sentence, and names double exactly once - to say the course does not use
it, and why. Lesson 3's division demo is 7.0f / 2, in the sample's own
printed label as well as the prose. Lesson 14's gravity try-it keeps its
whole-number step (try 2) to show the suffix is for decimals only.
Lesson 18 introduces MathF alongside Unity's Mathf. Lesson 19's try-it
needs 1.1f and 1.3f - without the suffix those lines no longer compile.

One golden changed: s03a-division.out.txt, because the sample's label
string changed, not because float printed differently. Every frame
golden - 20 sample, 9 challenge - reproduces bit-for-bit: the games use
whole numbers and one 0.5, all exact in float, and the renderer's own
maths is untouched.

+++ content/challenges/c14-bouncing-ball.solution.cs
@@ -1,8 +1,8 @@
-double x = 320;
-double y = 180;
-double speedX = 4;
-double speedY = 3;
-double radius = 20;
+float x = 320;
+float y = 180;
+float speedX = 4;
+float speedY = 3;
+float radius = 20;
 
 void Setup()
 {
